crs/src/requirements.txt
requests==2.32.3
PyYAML==6.0.1
GitPython==3.1.43
langchain==0.2.5
langchain-openai==0.1.8
langgraph==0.1.1
StrEnum==0.4.15
